[FIX] Multiple InsertChild
The current implementation does not extend the Child slice of the element, so if we call several times the `InsertChild()` method, the end of the Child slice is deleted as described in the documentation of the `copy` builtin (copied elements are the min of the size of the 2 slices).
This PR increases the size of the Child slice to make sure that the insertion doesn't remove any other child.