add support for absolute path queries.
Paths that are prefixed with '/' or '//' are now handled.  If a
path is prefixed with '//', it will perform an all-descendant
search starting from the root.  If a path is prefixed with a
single '/', it will perform a search starting from the root.

THIS IS A BREAKING CHANGE FOR '//' QUERIES.

Previously, a path that was prefixed with '//' would begin from
the current element, not from the root. To retain the old
behavior, the path must be prefixed with './/' instead of '//'.