Improve indent APIs
Introduce the 'Unindent' function to remove indentation from a
document. This is a convenience function that replaces
'Indent(etree.NoIndent)'.

Add an 'IndentWithSettings' function to Document. This function
allows you more control over the indentation algorithm, which is
configured with the new 'IndentSettings' struct. UseCRLF and
PreserveLeafWhitespace were moved from Document to this new
InputSettings struct. (Document.UseCRLF will continue to work,
but it is deprecated since it has existed in the API for a long
time. Document.PreserveLeafWhitespace has been totally removed
since it was public for only a couple hours.)