Default CharsetReader improvements
When nil, the ReadSettings struct's CharsetReader field now causes
the XML decoder to use a "pass-though" charset converter, passing
the reader's data through without modification.

This was already the default behavior when creating a new etree
document with the NewDocument function, but now a default-
constructed ReadSettings struct will result in the same default
CharsetReader behavior.